Split document processing in separate files per document type
