+++ api/src/changeRequests/documentProcessing/processPostTagDto.ts
@@ -1,0 +1,93 @@
+import { ContentDto } from "../../dto/ContentDto";
+import { PostDto } from "../../dto/PostDto";
+import { TagDto } from "../../dto/TagDto";
+import { DbService } from "../../db/db.service";
+import { DocType, Uuid } from "../../enums";
+import { processImage } from "../../s3/s3.imagehandling";
+import { S3Service } from "../../s3/s3.service";
+
+/**
+ * Process Post / Tag DTO
+ * @param doc
+ * @param db
+ */
+export default async function processPostTagDto(
+    doc: PostDto | TagDto,
+    prevDoc: PostDto | TagDto,
+    db: DbService,
+    s3: S3Service,
+) {
+    // Cascade delete for Post and Tag documents to content documents
+    if (doc.deleteReq) {
+        const contentDocs = await db.getContentByParentId(doc._id);
+        for (const contentDoc of contentDocs.docs) {
+            await db.deleteDoc(contentDoc._id);
+        }
+
+        return; // no need to process further
+    }
+
+    // Process image uploads
+    if ((doc as PostDto).imageData) {
+        await processImage(doc.imageData, prevDoc?.imageData, s3);
+        delete (doc as any).image; // Remove the legacy image field
+    }
+
+    // Get content documents that are children of the Post / Tag document
+    // and copy essential properties from the Post / Tag document to the child content document
+    const contentDocs = await db.getContentByParentId(doc._id);
+    for (const contentDoc of contentDocs.docs) {
+        contentDoc.memberOf = doc.memberOf;
+        contentDoc.parentTags = doc.tags;
+        contentDoc.parentImageData = doc.imageData;
+
+        if (doc.type == DocType.Post) {
+            contentDoc.parentPostType = (doc as PostDto).postType;
+        }
+
+        if (doc.type == DocType.Tag) {
+            contentDoc.parentTagType = (doc as TagDto).tagType;
+            contentDoc.parentPinned = (doc as TagDto).pinned;
+        }
+
+        contentDoc.parentPublishDateVisible = doc.publishDateVisible;
+        await db.upsertDoc(contentDoc);
+    }
+
+    // tag caching to the taggedDocs / parentTaggedDocs property of tag / content documents. This is done to improve client query performance.
+    const addedTags = prevDoc ? doc.tags.filter((tag) => !prevDoc.tags.includes(tag)) : [];
+    const removedTags = prevDoc ? prevDoc.tags.filter((tag) => !doc.tags.includes(tag)) : [];
+    const changedTags = addedTags
+        .concat(removedTags)
+        .filter((tag, index, self) => self.indexOf(tag) === index);
+    const tagDocs = changedTags.length ? (await db.getDocs(changedTags, [DocType.Tag])).docs : [];
+    const tagDocsContent = changedTags.length
+        ? (await db.getContentByParentId(changedTags)).docs
+        : [];
+    const updatedDocs = tagDocs.concat(tagDocsContent);
+
+    for (const d of updatedDocs) {
+        let taggedDocsArray: Uuid[];
+        let tagId: Uuid;
+        if (d.type == DocType.Tag) {
+            const tag = d as TagDto;
+            tag.taggedDocs = tag.taggedDocs || [];
+            taggedDocsArray = tag.taggedDocs;
+            tagId = tag._id;
+        } else {
+            const content = d as ContentDto;
+            content.parentTaggedDocs = content.parentTaggedDocs || [];
+            taggedDocsArray = content.parentTaggedDocs;
+            tagId = content.parentId;
+        }
+
+        if (addedTags.includes(tagId)) taggedDocsArray.push(doc._id);
+
+        if (removedTags.includes(tagId)) {
+            const index = taggedDocsArray.indexOf(doc._id);
+            if (index > -1) taggedDocsArray.splice(index, 1);
+        }
+
+        await db.upsertDoc(d);
+    }
+}
