+++ api/src/changeRequests/processChangeRequest.ts
@@ -2,14 +2,16 @@ import { validateChangeRequest } from "./validateChangeRequest";
 import { DbService, DbUpsertResult } from "../db/db.service";
 import { ChangeReqDto } from "../dto/ChangeReqDto";
 import { DocType, Uuid } from "../enums";
-import { validateSlug } from "./validateSlug";
-import { processImage } from "../s3/s3.imagehandling";
 import { S3Service } from "../s3/s3.service";
 import { PostDto } from "../dto/PostDto";
 import { TagDto } from "../dto/TagDto";
 import { ContentDto } from "../dto/ContentDto";
 import { LanguageDto } from "../dto/LanguageDto";
 import { isEqualDoc } from "../util/isEqualDoc";
+import { _baseDto } from "src/dto/_baseDto";
+import processPostTagDto from "./documentProcessing/processPostTagDto";
+import processContentDto from "./documentProcessing/processContentDto";
+import processLanguageDto from "./documentProcessing/processLanguageDto";
 
 export async function processChangeRequest(
     userId: string,
@@ -25,20 +27,14 @@ export async function processChangeRequest(
         throw new Error(validationResult.error);
     }
 
-    // Reject if a language document is deleted with the default language field set
-    if (
-        changeRequest.doc.type === DocType.Language &&
-        changeRequest.doc.deleteReq &&
-        (changeRequest.doc as LanguageDto).default
-    ) {
-        throw new Error("Cannot delete the default language document");
-    }
-
     const doc = validationResult.validatedData;
-    const prevDoc = await db.getDoc(doc._id);
+    const prevDocQuery = await db.getDoc(doc._id);
+    const prevDoc: _baseDto | undefined = prevDocQuery.docs?.length
+        ? prevDocQuery.docs[0]
+        : undefined;
 
     // Check if the document has changed
-    if (prevDoc.docs && prevDoc.docs.length && isEqualDoc(doc, prevDoc.docs[0])) {
+    if (isEqualDoc(doc, prevDoc)) {
         return {
             ok: true,
             message: "Document is identical to the one in the database",
@@ -48,159 +44,14 @@ export async function processChangeRequest(
     // insert user id into the change request document, so that we can keep a record of who made the change
     doc.updatedBy = userId;
 
-    if (!doc.deleteReq) {
-        // Validate slug
-        if (doc.type == DocType.Content) {
-            doc.slug = await validateSlug(doc.slug, doc._id, db);
-        }
+    const docProcessMap = {
+        [DocType.Post]: () => processPostTagDto(doc as PostDto, prevDoc as PostDto, db, s3),
+        [DocType.Tag]: () => processPostTagDto(doc as TagDto, prevDoc as TagDto, db, s3),
+        [DocType.Content]: () => processContentDto(doc as ContentDto, db),
+        [DocType.Language]: () => processLanguageDto(doc as LanguageDto, db),
+    };
 
-        if (doc.type == DocType.Content) {
-            const contentDoc = doc as ContentDto;
-            const parentQuery = await db.getDoc(doc.parentId);
-            const parentDoc: PostDto | TagDto | undefined =
-                parentQuery.docs.length > 0 ? parentQuery.docs[0] : undefined;
-
-            // Copy essential properties from Post / Tag documents to Content documents
-            if (parentDoc) {
-                const contentDoc = doc as ContentDto;
-                contentDoc.memberOf = parentDoc.memberOf;
-                contentDoc.parentTags = parentDoc.tags;
-                contentDoc.parentImageData = parentDoc.imageData;
-
-                if (parentDoc.type == DocType.Post) {
-                    contentDoc.parentPostType = (parentDoc as PostDto).postType;
-                }
-
-                if (parentDoc.type == DocType.Tag) {
-                    contentDoc.parentTagType = (parentDoc as TagDto).tagType;
-                    contentDoc.parentPinned = (parentDoc as TagDto).pinned;
-                }
-                contentDoc.parentPublishDateVisible = parentDoc.publishDateVisible;
-            }
-
-            // Find all available translations, and add them to the content document's availableTranslations property
-            const translations = await db.getContentByParentId(parentDoc._id);
-            const uniqueLanguages = new Set<Uuid>(translations.docs.map((doc) => doc.language));
-            uniqueLanguages.add(contentDoc.language);
-            const availableTranslations = Array.from(uniqueLanguages);
-            contentDoc.availableTranslations = availableTranslations;
-
-            // Update all translations with the new list of available translations
-            for (const translation of translations.docs) {
-                translation.availableTranslations = availableTranslations;
-                await db.upsertDoc(translation);
-            }
-        }
-
-        if (doc.type == DocType.Post || doc.type == DocType.Tag) {
-            // Process image uploads
-            if ((doc as PostDto).imageData) {
-                const prevImageData =
-                    prevDoc.docs.length > 0 ? prevDoc.docs[0].imageData : undefined;
-                await processImage(doc.imageData, prevImageData, s3);
-                delete doc.image; // Remove the legacy image field
-            }
-
-            // Get content documents that are children of the Post / Tag document
-            // and copy essential properties from the Post / Tag document to the child content document
-            const contentDocs = await db.getContentByParentId(doc._id);
-            for (const contentDoc of contentDocs.docs) {
-                contentDoc.memberOf = doc.memberOf;
-                contentDoc.parentTags = doc.tags;
-                contentDoc.parentImageData = doc.imageData;
-
-                if (doc.type == DocType.Post) {
-                    contentDoc.parentPostType = (doc as PostDto).postType;
-                }
-
-                if (doc.type == DocType.Tag) {
-                    contentDoc.parentTagType = (doc as TagDto).tagType;
-                    contentDoc.parentPinned = (doc as TagDto).pinned;
-                }
-
-                contentDoc.parentPublishDateVisible = doc.publishDateVisible;
-                await db.upsertDoc(contentDoc);
-            }
-
-            // tag caching to the taggedDocs / parentTaggedDocs property of tag / content documents. This is done to improve client query performance.
-            const prevTags = prevDoc.docs.length ? (prevDoc.docs[0] as PostDto | TagDto).tags : [];
-            const addedTags = (doc as PostDto | TagDto).tags.filter(
-                (tag) => !prevTags.includes(tag),
-            );
-            const removedTags = prevTags.filter(
-                (tag) => !(doc as PostDto | TagDto).tags.includes(tag),
-            );
-            const changedTags = addedTags
-                .concat(removedTags)
-                .filter((tag, index, self) => self.indexOf(tag) === index);
-            const tagDocs = changedTags.length
-                ? (await db.getDocs(changedTags, [DocType.Tag])).docs
-                : [];
-            const tagDocsContent = changedTags.length
-                ? (await db.getContentByParentId(changedTags)).docs
-                : [];
-            const updatedDocs = tagDocs.concat(tagDocsContent);
-
-            for (const d of updatedDocs) {
-                let taggedDocsArray: Uuid[];
-                let tagId: Uuid;
-                if (d.type == DocType.Tag) {
-                    const tag = d as TagDto;
-                    tag.taggedDocs = tag.taggedDocs || [];
-                    taggedDocsArray = tag.taggedDocs;
-                    tagId = tag._id;
-                } else {
-                    const content = d as ContentDto;
-                    content.parentTaggedDocs = content.parentTaggedDocs || [];
-                    taggedDocsArray = content.parentTaggedDocs;
-                    tagId = content.parentId;
-                }
-
-                if (addedTags.includes(tagId)) taggedDocsArray.push(doc._id);
-
-                if (removedTags.includes(tagId)) {
-                    const index = taggedDocsArray.indexOf(doc._id);
-                    if (index > -1) taggedDocsArray.splice(index, 1);
-                }
-
-                await db.upsertDoc(d);
-            }
-        }
-
-        if (doc.type === DocType.Language) {
-            const langDoc = doc as LanguageDto;
-
-            if (langDoc.default == 1) {
-                const languageDocs = await db.getDocsByType(DocType.Language);
-
-                languageDocs.docs.forEach(async (doc: LanguageDto) => {
-                    if (langDoc._id == doc._id) return;
-                    await db.upsertDoc({ ...doc, default: 0 });
-                });
-            }
-        }
-    }
-
-    // Cascade delete for Post and Tag documents
-    if (doc.deleteReq && (doc.type == DocType.Post || doc.type == DocType.Tag)) {
-        const contentDocs = await db.getContentByParentId(doc._id);
-        for (const contentDoc of contentDocs.docs) {
-            await db.deleteDoc(contentDoc._id);
-        }
-    }
-
-    if (doc.deleteReq && doc.type == DocType.Content) {
-        const parentQuery = await db.getDoc(doc.parentId);
-        const parentDoc: PostDto | TagDto | undefined = parentQuery.docs.length
-            ? parentQuery.docs[0]
-            : undefined;
-
-        // We still need to replicate the parent document's memberOf property to the content document, as this is not (and should not be) done in the CMS.
-        if (parentDoc) {
-            const contentDoc = doc as ContentDto;
-            contentDoc.memberOf = parentDoc.memberOf;
-        }
-    }
+    docProcessMap[doc.type] && (await docProcessMap[doc.type]());
 
     // Insert / update the document in the database
     const upsertResult = await db.upsertDoc(doc);
